pyproject.toml
[tool.poetry]
name = "project-lazy-boy"
version = "0.1.0"
description = ""
authors = ["ma2za <mazzapaolo2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
httpx = { extras = ["http2"], version = "^0.25.0" }
beautifulsoup4 = "^4.12.2"
redis = "^5.0.1"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.scraper.dependencies]
scipy = "^1.11.3"
openai = "^0.28.1"
torch = { version = "2.1.0+cpu", source = "torch" }
transformers = {extras = ["torch"], version = "^4.34.1"}


[tool.poetry.group.career-scraping.dependencies]
trafilatura = "^1.6.2"
langchain = "^0.0.320"
openai = "^0.28.1"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/"
default = false
secondary = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
